requirements.txt
asgiref==3.8.1
click==8.1.7
cmake==3.25.2
Django==5.1
djangorestframework==3.15.2
dlib==19.24.2
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==1.24.2
opencv-python==4.10.0.84
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
